Add 92 tests for Diagnostics+Text remaining (Task 35); fix Rune/Encoding headers; total 1824 tests
- New: DiagnosticsRemainingTests.cpp (29 tests) — DebuggerBrowsableState enum,
  DebuggerBrowsable/DisplayAttribute, StackFrame, StackTrace, UnreachableException,
  marker attrs (Hidden/StepThrough/NonUserCode)
- New: TextRemainingTests.cpp (63 tests) — NormalizationForm, CompositeFormat,
  Rune (full coverage incl. utf8/utf16 seq lengths, IsLetter/Digit/Whitespace,
  ToUpper/ToLower, ReplacementChar), UTF7/UTF32/Latin1Encoding, Encoder/Decoder,
  Regex/Match/MatchCollection
- Fix Rune.hpp: ReplacementChar{0xFFFD} → {uint32_t(0xFFFD)} to resolve ambiguous
  ctor overload between Rune(uint32_t) and Rune(char)
- Fix Encoding.hpp: add virtual getCodePageProperty() so UTF32/Latin1Encoding
  overrides compile without error

+++ include/System/Text/Rune.hpp
@@ -88,6 +88,6 @@ namespace System::Text {
         bool operator>=(const Rune& o) const { return value_ >= o.value_; }
     };
 
-    inline const Rune Rune::ReplacementChar{0xFFFD};
+    inline const Rune Rune::ReplacementChar{uint32_t(0xFFFD)};
 
 } // namespace System::Text
